*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT PI/win32/org/eclipse/swt/internal/win32/OS.java
@@ -2491,7 +2491,7 @@ public static final native boolean RoundRect (int hdc, int nLeftRect, int nTopRe
 public static final native int SaveDC(int hdc);
 public static final native boolean ScreenToClient (int hWnd, POINT lpPoint);
 public static final native int ScriptBreak(char[] pwcChars, int cChars, SCRIPT_ANALYSIS psa, int psla);
-public static final native int ScriptCacheGetHeight(int hdc, int psc, long[] tmHeight);
+public static final native int ScriptCacheGetHeight(int hdc, int psc, int[] tmHeight);
 public static final native int ScriptCPtoX(int iCP, boolean fTrailing, int cChars, int cGlyphs, int pwLogClust, int psva, int piAdvance, SCRIPT_ANALYSIS psa, int[] piX);
 public static final native int ScriptFreeCache(int psc);
 public static final native int ScriptGetLogicalWidths (SCRIPT_ANALYSIS psa, int cChars, int cGlyphs, int piGlyphWidth, int pwLogClust, int psva, int[] piDx);
